[refactor] linear tick locator (#63)
This change simplifies the linear tick locator a bit. Importantly, it's
now possible to use non-integer tick-distances like `calc.pi` while
keeping all ticks multiples of that distance (i.e., 0 is always a tick).

Changing the tick locator is a bit delicate because it's hard to ensure
that all special cases with degrading numeric precision still work but
we have many tests and they still pass.