[add] `unit` to linear tick locator (#64)
This makes it possible to specify a unit other than 1. This is for
example useful to generate ticks that are multiples of $\pi$ (or any
other real number). This is however more powerful than setting the tick
distance because, the tick step is still smartly selected depending on
the range − just based on $\pi$ multiples.

At the same time, we introduce a `suffix` for the linear tick formatter
that allows to add a suffix to each number, except when the number is 0,
1, or −1. Also, the linear tick formatter is aware of the `unit` that
has been set with the locator.


## Example 
```typ
#lq.diagram(
  xlim: (-2*calc.pi, 2*calc.pi),

  xaxis: (
    locate-ticks: lq.locate-ticks-linear.with(unit: calc.pi),
    format-ticks: lq.format-ticks-linear.with(suffix: $pi$)
  )
)
```

![image](https://github.com/user-attachments/assets/50019ead-b6bd-4b96-b10f-66758a555084)