[replace] table with grid in `legend` (#78)
This replaces the usage of `table` with `grid` for layouting the
items of a `legend`.

Semantically one could argue for both grid or table to be appropriate
for a legend. However, tables are common to be styled globally in a
document. It should be extremely rare that this should also affect the
legend table, leading to undesired results, see #77. Grids on the other
hand are much less likely to be styled globally.


⚠️ This will require users to change rules of the form
```typ
#show lq.selector(lq.legend): set table(..)
```
to 
```typ
#show lq.selector(lq.legend): set grid(..)
```


Closes #77.