[add] allow header names as converter names (#86)
While playing around with the `load-txt` function, I found it very
unintuitive that you can't use the column names, as defined by the
header, to define converters. This PR amends that while still retaining
the column index based converter definitions when headers are enabled.

Please note that I didn't go through the long-winded route of opening an
issue for this, as I assumed that this was too small of a change to
warrant the paper trail.

---------

Co-authored-by: Mc-Zen <52877387+Mc-Zen@users.noreply.github.com>