[add] Allow relative and ratio inputs to `diagram.width` and `diagram.height` (#93)
## Details
The dimensions of the _data area_ (confined by the diagram axes) are
computed by the following procedure:
1. First, the available size is retrieved via
[`layout`](https://typst.app/docs/reference/layout/layout/) and the
total target size of the diagram is computed.
2. Then, a rough estimate for the data area is guessed (subtracting the
space needed for axes, titles, legend etc.).
3. A first layout attempt is made with the provisional data area
dimensions. For this, the bounds of several diagram elements need to be
computed. This step includes:
- Axes tick generation and measurement of ticks, labels, and
attachments.
   - Layouting and measuring the title. 
- Generating, positioning, and measuring the legend, in case it juts out
of the data area.
   - Measuring `lq.place` instances that stick out of the data area. 
- From these results, computing the margins needed to the left, top,
right, and bottom of the data area.
4. The width/height remaining for the data area (equivalent to the
length of the axes) is computed based on the results of step 3.
5. Step 3 is repeated with the new dimensions of the data area. Now we
have a much better estimate for how long the axes will be. These
dimensions might deviate significantly from the first rough guess from
step 2 and we need to give the tick locators a chance to adjust to the
new lengths (the optimal number of ticks that the locators can place on
the axes depends on the available axis length).
6. Again, the remaining width/height for the data area is computed. We
can safely assume that the axis lengths did not change significantly
this time and thus we fix the located ticks from step 5 to be sure that
the margins do not change again.

Now, we have the final dimensions for the data area and follow the
normal plotting procedure. The only difference is that the axis ticks
have already been computed based on the dimensions computed in step 4.

I would argue that two iterations should be enough for the dimensions to
stabilize sufficiently for the ticks to be spaced appropriately.



## Performance
Performance-wise, things are looking good. There is some overhead due to
the repeated layouting and measuring iterations but it is not
significant and virtually constant in the number of plots per diagram,.
Also, this overhead is only paid when relative sizing is actually used
and not when fixed diagram dimensions are used.