[fix] rounding of tick-distance (#125)
This fixes a rounding issue with the linear tick locator. To compensate
issues with degrading floating point precision, the tick distance needs
to be rounded. This PR increases the number of digits to round.

This is in order to allow tick distances like 1/3 which would otherwise
be rounded to something like 0.33333 only instead of 0.333333333333
leading to serious problems for example with the fractional tick
formatter.