[fix] spacing with labels longer than the diagram (#139)
This is an amendment of #67. We're doing the job properly now and just
wrap tick labels in an infinite box. With this approach, the computation
of bounds also works when a transformational show rule like a rotation
of the labels transforms them into a new size.

This applies to a range of elements:
- title
- label
- tick-label

This PR also disallows 2d aligments for `title.position`. Only left,
right, top, and bottom are valid.